+++ bin/main.dart
@@ -1,5 +1,20 @@
-import 'package:nylo_framework/metro/metro.dart' as metro;
+import 'dart:io';
 
-void main(List<String> arguments) {
-  metro.commands(arguments);
+import 'package:nylo_framework/metro/metro.dart' as metro_cli;
+import 'package:nylo_support/metro/metro_console.dart';
+
+void main(List<String> arguments) async {
+  var time = Stopwatch();
+  bool hasDebugFlag = arguments.contains("--debug");
+  if (hasDebugFlag == true) {
+    time.start();
+    arguments.removeWhere((element) => element == "--debug");
+  }
+  await metro_cli.commands(arguments);
+  if (hasDebugFlag == true) {
+    time.stop();
+    MetroConsole.writeInBlack(
+        'Time: ${time.elapsed.inMilliseconds} Milliseconds');
+  }
+  exit(0);
 }