+++ lib/metro/metro.dart
@@ -3,7 +3,10 @@ library metro;
 import 'package:collection/collection.dart' show IterableExtension;
 import 'package:flutter_launcher_icons/main.dart' as launcherIcons;
 import 'package:args/args.dart';
-import 'package:nylo_framework/metro/constants/strings.dart';
+import 'package:nylo_framework/metro/menu.dart';
+import 'package:nylo_support/metro/constants/strings.dart';
+import 'package:nylo_support/metro/metro_console.dart';
+import 'package:nylo_support/metro/metro_service.dart';
 import 'package:nylo_framework/metro/models/ny_command.dart';
 import 'package:nylo_framework/metro/stubs/controller_stub.dart';
 import 'package:nylo_framework/metro/stubs/model_stub.dart';
@@ -13,131 +16,77 @@ import 'package:nylo_framework/metro/stubs/widget_stateful_stub.dart';
 import 'package:nylo_framework/metro/stubs/widget_stateless_stub.dart';
 import 'dart:io';
 
-List<NyCommand> _allProjectCommands = [
-  NyCommand(name: "init", options: 1, arguments: ["-h"], action: _projectInit),
-];
-
-List<NyCommand> _allMakeCommands = [
+List<NyCommand> _allCommands = [
+  NyCommand(
+      name: "init",
+      options: 1,
+      arguments: ["-h"],
+      category: "project",
+      action: _projectInit),
   NyCommand(
       name: "controller",
       options: 1,
       arguments: ["-h", "-f"],
+      category: "make",
       action: _makeController),
   NyCommand(
       name: "model",
       options: 1,
       arguments: ["-h", "-f", "-s"],
+      category: "make",
       action: _makeModel),
   NyCommand(
-      name: "page", options: 1, arguments: ["-h", "-f"], action: _makePage),
+      name: "page",
+      options: 1,
+      category: "make",
+      arguments: ["-h", "-f"],
+      action: _makePage),
   NyCommand(
       name: "stateful_widget",
       options: 1,
       arguments: ["-h", "-f"],
+      category: "make",
       action: _makeStatefulWidget),
   NyCommand(
       name: "stateless_widget",
       options: 1,
       arguments: ["-h", "-f"],
+      category: "make",
       action: _makeStatelessWidget),
-];
-
-List<NyCommand> _allAppIconsCommands = [
   NyCommand(
-      name: "build", options: 1, arguments: ["-h"], action: _makeAppIcons),
+      name: "build",
+      options: 1,
+      arguments: ["-h"],
+      category: "appicons",
+      action: _makeAppIcons),
 ];
 
 Future<void> commands(List<String> arguments) async {
-  if (arguments.length == 0) {
-    _handleMenu();
-    exit(0);
+  if (arguments.isEmpty) {
+    MetroConsole.writeInBlack(metroMenu);
+    return;
   }
 
   List<String> argumentSplit = arguments[0].split(":");
 
   if (argumentSplit.length == 0 || argumentSplit.length <= 1) {
-    writeInBlack('Invalid arguments ' + arguments.toString());
+    MetroConsole.writeInBlack('Invalid arguments ' + arguments.toString());
     exit(2);
   }
 
   String type = argumentSplit[0];
   String action = argumentSplit[1];
 
-  NyCommand? nyCommand;
-  switch (type) {
-    case "project":
-      {
-        nyCommand = _allProjectCommands
-            .firstWhereOrNull((command) => command.name == action);
-        break;
-      }
-    case "make":
-      {
-        nyCommand = _allMakeCommands
-            .firstWhereOrNull((command) => command.name == action);
-        break;
-      }
-    case "appicons":
-      {
-        nyCommand = _allAppIconsCommands
-            .firstWhereOrNull((command) => command.name == action);
-        break;
-      }
-    case "plugin":
-      {
-        nyCommand = _allAppIconsCommands
-            .firstWhereOrNull((command) => command.name == action);
-        break;
-      }
-    default:
-      {}
-  }
+  NyCommand? nyCommand = _allCommands.firstWhereOrNull(
+      (command) => type == command.category && command.name == action);
 
   if (nyCommand == null) {
-    writeInBlack('Invalid arguments ' + arguments.toString());
-    return;
+    MetroConsole.writeInBlack('Invalid arguments ' + arguments.toString());
+    exit(1);
   }
 
   arguments.removeAt(0);
-  nyCommand.action!(arguments);
-}
-
-_handleMenu() {
-  writeInBlack(
-      'Metro - Nylo\'s Companion to Build Flutter apps by Anthony Gordon');
-  writeInBlack('Version: 1.0.0');
-
-  writeInBlack('');
-
-  writeInBlack('Usage: ');
-  writeInBlack('    command [options] [arguments]');
-
-  writeInBlack('');
-  writeInBlack('Options');
-  writeInBlack('    -h');
-
-  writeInBlack('');
-
-  writeInBlack('All commands:');
-
-  _writeInGreen(' project');
-  _allProjectCommands.forEach((command) {
-    writeInBlack('  project:' + command.name!);
-  });
-
-  writeInBlack('');
-
-  _writeInGreen(' make');
-  _allMakeCommands.forEach((command) {
-    writeInBlack('  make:' + command.name!);
-  });
-
-  writeInBlack('');
-
-  _writeInGreen(' appicons');
-  _allAppIconsCommands.forEach((command) {
-    writeInBlack('  appicons:' + command.name!);
-  });
+  await nyCommand.action!(arguments);
 }
 
 _projectInit(List<String> arguments) async {
@@ -148,19 +97,16 @@ _projectInit(List<String> arguments) async {
 
   final ArgResults argResults = parser.parse(arguments);
 
-  if (argResults[helpFlag]) {
-    writeInBlack('Help - Initializes the Nylo project');
-    writeInBlack(parser.usage);
-    exit(0);
+  _checkHelpFlag(argResults[helpFlag], parser.usage);
+
+  if (await MetroService.hasFile(".env") == false) {
+    final File envExample = File('.env-example');
+    final File env = File('.env');
+    await env.writeAsString(await envExample.readAsString());
+    MetroConsole.writeInGreen('.env file add from .env-example 🎉');
   }
 
-  final File envExample = File('.env-example');
-  final File env = File('.env');
-
-  await env.writeAsString(await envExample.readAsString());
-
-  _writeInGreen('Project initialized, create something great 🎉');
-  exit(0);
+  MetroConsole.writeInGreen('Project initialized, create something great 🎉');
 }
 
 _makeStatefulWidget(List<String> arguments) async {
@@ -178,32 +124,17 @@ _makeStatefulWidget(List<String> arguments) async {
   final ArgResults argResults = parser.parse(arguments);
 
   bool? hasForceFlag = argResults[forceFlag];
-  bool hasHelpFlag = argResults[helpFlag];
 
-  if (hasHelpFlag) {
-    writeInBlack(parser.usage);
-    exit(0);
-  }
+  _checkHelpFlag(argResults[helpFlag], parser.usage);
 
   String widgetName =
       argResults.arguments.first.replaceAll(RegExp(r'(_?widget)'), "");
 
-  String path = '$widgetFolder/${widgetName.toLowerCase()}_widget.dart';
+  String stubStatefulWidget = widgetStatefulStub(_parseToPascal(widgetName));
+  await MetroService.makeStatefulWidget(widgetName, stubStatefulWidget,
+      forceCreate: hasForceFlag ?? false);
 
-  if (await File(path).exists() && hasForceFlag == false) {
-    _writeInRed(widgetName + '_widget already exists');
-    exit(0);
-  }
-
-  final File file = File(path);
-
-  await _makeDirectory(widgetFolder);
-
-  await file.writeAsString(widgetStatefulStub(_parseToPascal(widgetName)));
-
-  _writeInGreen(widgetName + '_widget created 🎉');
-
-  exit(0);
+  MetroConsole.writeInGreen(widgetName + '_widget created 🎉');
 }
 
 _makeStatelessWidget(List<String> arguments) async {
@@ -221,32 +152,20 @@ _makeStatelessWidget(List<String> arguments) async {
   final ArgResults argResults = parser.parse(arguments);
 
   bool? hasForceFlag = argResults[forceFlag];
-  bool hasHelpFlag = argResults[helpFlag];
 
-  if (hasHelpFlag) {
-    writeInBlack(parser.usage);
-    exit(0);
-  }
+  _checkHelpFlag(argResults[helpFlag], parser.usage);
+
+  _checkArguments(arguments,
+      'You are missing the \'name\' of the widget that you want to create.\ne.g. make:stateless_widget my_new_widget');
 
   String widgetName =
       argResults.arguments.first.replaceAll(RegExp(r'(_?widget)'), "");
 
-  String path = '$widgetFolder/${widgetName.toLowerCase()}_widget.dart';
+  String stubStatelessWidget = widgetStatelessStub(_parseToPascal(widgetName));
+  await MetroService.makeStatelessWidget(widgetName, stubStatelessWidget,
+      forceCreate: hasForceFlag ?? false);
 
-  if (await File(path).exists() && hasForceFlag == false) {
-    _writeInRed(widgetName + '_widget already exists');
-    exit(0);
-  }
-
-  final File file = File(path);
-
-  await _makeDirectory(widgetFolder);
-
-  await file.writeAsString(widgetStatelessStub(_parseToPascal(widgetName)));
-
-  _writeInGreen(widgetName + '_widget created 🎉');
-
-  exit(0);
+  MetroConsole.writeInGreen(widgetName + '_widget created 🎉');
 }
 
 _makeAppIcons(List<String> arguments) async {
@@ -259,16 +178,11 @@ _makeAppIcons(List<String> arguments) async {
 
   final ArgResults argResults = parser.parse(arguments);
 
-  if (argResults[helpFlag]) {
-    writeInBlack('Help - Creates App Icons in Nylo');
-    writeInBlack(parser.usage);
-    exit(0);
-  }
+  _checkHelpFlag(argResults[helpFlag], parser.usage);
 
   launcherIcons.createIconsFromArguments(arguments);
 
-  _writeInGreen('App icons created 🎉');
-  exit(0);
+  MetroConsole.writeInGreen('App icons created 🎉');
 }
 
 _makeController(List<String> arguments) async {
@@ -284,37 +198,21 @@ _makeController(List<String> arguments) async {
       negatable: false);
 
   final ArgResults argResults = parser.parse(arguments);
-
-  if (argResults.arguments.length == 0) {
-    writeInBlack(parser.usage);
-    exit(0);
-  }
+  _checkArguments(arguments, parser.usage);
 
   bool? hasForceFlag = argResults[forceFlag];
-  bool hasHelpFlag = argResults[helpFlag];
-
-  if (hasHelpFlag) {
-    writeInBlack(parser.usage);
-    exit(0);
-  }
+  _checkHelpFlag(argResults[helpFlag], parser.usage);
 
   String className =
       argResults.arguments.first.replaceAll(RegExp(r'(_?controller)'), "");
-  String path = '$controllerFolder/${className.toLowerCase()}_controller.dart';
 
-  if (await File(path).exists() && hasForceFlag == false) {
-    _writeInRed(className + ' already exists');
-    exit(0);
-  }
+  String stubController =
+      controllerStub(controllerName: _parseToPascal(className));
 
-  String controllerName = _parseToPascal(className);
-  String strController = controllerStub(controllerName: controllerName);
+  await MetroService.makeController(className, stubController,
+      forceCreate: hasForceFlag ?? false);
 
-  await _writeToFilePath(path, strController);
-
-  _writeInGreen(className + '_controller created 🎉');
-
-  exit(0);
+  MetroConsole.writeInGreen(className + '_controller created 🎉');
 }
 
 _makeModel(List<String> arguments) async {
@@ -333,137 +231,73 @@ _makeModel(List<String> arguments) async {
 
   final ArgResults argResults = parser.parse(arguments);
 
-  if (argResults.arguments.length == 0) {
-    writeInBlack(parser.usage);
-    exit(0);
-  }
+  _checkArguments(argResults.arguments, parser.usage);
 
   bool? hasForceFlag = argResults[forceFlag];
-  bool hasHelpFlag = argResults[helpFlag];
   bool? hasStorableFlag = argResults[storableFlag];
 
-  if (hasHelpFlag) {
-    writeInBlack(parser.usage);
-    exit(0);
-  }
+  _checkHelpFlag(argResults[helpFlag], parser.usage);
 
-  String firstArg = argResults.arguments.first;
-  String path = '$modelFolder/${firstArg.toLowerCase()}.dart';
-
-  if (await File(path).exists() && hasForceFlag == false) {
-    _writeInRed(firstArg + ' already exists');
-    exit(0);
-  }
-
-  final File file = File(path);
-
-  String modelName = _parseToPascal(firstArg);
-  String strModel =
+  String className = argResults.arguments.first;
+  String modelName = _parseToPascal(className);
+  String stubModel =
       modelStub(modelName: modelName, isStorable: hasStorableFlag);
+  await MetroService.makeModel(className, stubModel,
+      forceCreate: hasForceFlag ?? false);
 
-  await _makeDirectory(modelFolder);
-
-  await file.writeAsString(strModel);
-
-  _writeInGreen(modelName + ' created 🎉');
-
-  exit(0);
+  MetroConsole.writeInGreen(modelName + ' created 🎉');
 }
 
 _makePage(List<String> arguments) async {
   final ArgParser parser = ArgParser(allowTrailingOptions: true);
 
-  parser.addFlag(helpFlag,
-      abbr: 'h',
-      help: 'Creates a new page widget for your project.',
-      negatable: false,
+  parser.addFlag(
+    helpFlag,
+    abbr: 'h',
+    help: 'Creates a new page widget for your project.',
+    negatable: false,
   );
-
-  bool shouldCreateController = false;
-  parser.addFlag(controllerFlag,
-      abbr: 'c',
-      help: 'Creates a new page with a controller',
-      negatable: false,
+  parser.addFlag(
+    controllerFlag,
+    abbr: 'c',
+    help: 'Creates a new page with a controller',
+    negatable: false,
   );
 
   final ArgResults argResults = parser.parse(arguments);
 
-  if (argResults[helpFlag]) {
-    writeInBlack('Help - Creates a new page in Nylo');
-    writeInBlack(parser.usage);
-    exit(0);
-  }
+  _checkHelpFlag(argResults[helpFlag], parser.usage);
 
+  bool shouldCreateController = false;
   if (argResults["controller"]) {
     shouldCreateController = true;
   }
 
-  if (argResults.arguments.first == " ") {
-    exit(0);
+  if (argResults.arguments.first.trim() == "") {
+    MetroConsole.writeInRed('You cannot create a page with an empty string');
+    exit(1);
   }
 
   String className =
       argResults.arguments.first.replaceAll(RegExp(r'(_?page)'), "");
 
-  String pathPage = '$pageFolder/${className.toLowerCase()}_page.dart';
-  String pathController =
-      '$controllerFolder/${className.toLowerCase()}_controller.dart';
   if (shouldCreateController) {
-    if (await File(pathPage).exists()) {
-      _writeInRed('${className}_page already exists');
-      return;
-    }
-
-    if (await File(pathController).exists()) {
-      _writeInRed('${className}_controller already exists');
-      return;
-    }
-
-    final File filePage = File(pathPage);
-
-    await _makeDirectory(controllerFolder);
-
-    await filePage.writeAsString(pageWithControllerStub(
+    String stubPageAndController = pageWithControllerStub(
         className: _parseToPascal(className),
-        importName: argResults.arguments.first.replaceAll("_page", "")));
+        importName: arguments.first.replaceAll("_page", ""));
+    await MetroService.makePage(className, stubPageAndController);
 
-    final File fileController = File(pathController);
-
-    String strController =
+    String stubController =
         controllerStub(controllerName: _parseToPascal(className));
+    await MetroService.makeController(className, stubController);
 
-    await fileController.writeAsString(strController);
-
-    _writeInGreen('${className}_page & ${className}_controller created 🎉');
-    exit(0);
+    MetroConsole.writeInGreen(
+        '${className}_page & ${className}_controller created 🎉');
+  } else {
+    String stubPage = pageStub(pageName: _parseToPascal(className));
+    await MetroService.makePage(className, stubPage);
+    MetroConsole.writeInGreen('${className}_page created 🎉');
   }
-
-  if (await File(pathPage).exists()) {
-    _writeInRed(argResults.arguments.first + ' already exists');
-    return;
-  }
-
-  final File file = File(pathPage);
-
-  await _makeDirectory(pageFolder);
-
-  await file.writeAsString(pageStub(pageName: _parseToPascal(className)));
-
-  _writeInGreen('${className}_page created 🎉');
-
-  exit(0);
-}
-
-_writeInGreen(String message) {
-  stdout.writeln('\x1B[92m' + message + '\x1B[0m');
-}
-
-_writeInRed(String message) {
-  stdout.writeln('\x1B[91m' + message + '\x1B[0m');
-}
-
-writeInBlack(String message) {
-  stdout.writeln(message);
 }
 
 String _parseToPascal(name) {
@@ -472,22 +306,16 @@ String _parseToPascal(name) {
   return map.join("");
 }
 
-_writeToFilePath(path, strFile) async {
-  final File file = File(path);
-
-  await file.writeAsString(strFile);
-}
-
-_makeDirectory(String path) async {
-  final File dirFolder = File(path);
-  if (!(await dirFolder.exists())) {
-    await Directory(path).create();
+_checkArguments(List<String> arguments, String usage) {
+  if (arguments.isEmpty) {
+    MetroConsole.writeInBlack(usage);
+    exit(1);
   }
 }
 
-String capitalize(String input) {
-  if (input.length == 0) {
-    return input;
+_checkHelpFlag(bool hasHelpFlag, String usage) {
+  if (hasHelpFlag) {
+    MetroConsole.writeInBlack(usage);
+    exit(0);
   }
-  return input[0].toUpperCase() + input.substring(1);
 }