+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [2.0.0] - 2021-09-10
+
+* Refactor nylo_framework.dart
+* Metro cli stability improvements
+* New NyTheme class for setting the Flutter theme
+* Null safety stubs
+* Pubspec.yaml dependency updates
+
 ## [1.1.0] - 2021-07-13
 
 * Null safety stubs